Fix compile warning in oic_string unit test.
Unit test was emitting a warning due to type mismatch of the constant.
Corrected it by explicitly using the proper type.

Reviewed-on: https://gerrit.iotivity.org/gerrit/1289
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>

+++ resource/c_common/oic_string/test/linux/oic_string_tests.cpp
@@ -271,7 +271,7 @@ TEST(StringTests, StrcatExtraRoom)
 
     char *result = OICStrcat(target, sizeof(target), source);
     EXPECT_EQ(target, result);
-    EXPECT_EQ(6, strlen(target));
+    EXPECT_EQ(static_cast<size_t>(6), strlen(target));
     EXPECT_STREQ("Orig12", target);
 
     for(size_t i = sizeof("Orig12"); i < sizeof(target); ++i)
